validation: Add BIP30Validate function
Move the BIP30 checks from ConnectBlock to a new BIP30Validate
Chainstate member function. This is a move-only change. The goal is to
move logic requiring access to the CCoinsViewCache out of ConnectBlock.
Over the next commits more utxo-set accessing logic is moved out of
ConnectBlock.

Callers of ConnectBlock now also need to call BIP30Validate before.

This is a part of a series of commits for removing access to the
CCoinsViewCache in consensus verification functions. The goal is to
allow calling verification functions with pre-fetched, or a user-defined
set of coins.