blockstorage: Replace BlockTreeDB with BlockTreeStore
This hooks up the newly introduced BlockTreeStore class to the actual
codebase. It also adds a migration function to migrate old leveldb block
indexes to the new format on startup.

The migration first reads from leveldb (blocks/index), and writes it to
a fresh BlockTreeStore in the same directory. Once done, the original
leveldb database is deleted.

On migration failure, the node shuts down again and prompts the user to
reindex. This is exercised in the added functional test. The functional
test uses version 28.2 to generate a legacy leveldb block tree db. This
version was chosen, since it is already required by other back-compat
tests (though other older versions would have worked too).

This commit is part of a series to replace the leveldb database
currently used for storing block indexes and headers with a flat file
storage. This is motivated by the kernel library, where the usage of
leveldb is a limiting factor to its future use cases.