blockstorage: Remove BlockTreeDB dead code
This is not called by anything anymore, so just remove it.

The max block file number and the last persisted block file info are now
always in lockstep, so remove the extra loop guarding against that.

This commit is part of a series to replace the leveldb database
currently used for storing block indexes and headers with a flat file
storage. This is motivated by the kernel library, where the usage of
leveldb is a limiting factor to its future use cases.